work towards regression test * refactor orchestrator to common package * sketch arc of whats needed in regression binary
